+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - calendar name column to form store
+- payment status, provider transaction id, and provider receipt number columns to form store
 
 ## [11.1.1] - 2026-07-28